Split unit test files according to separate modules for each class and outsource testdata
